Groundwork for Create Player (#21)
* Return correct status code

Kickstart the players endpoint by returning the correct status code on
player creation.

* Return received user

Start by returning the received user since it has the same json shape as
the response we want when we implement persistence in the database.

* Implement Player.create()

When we have a database this method will create a row in the players
table and return the associated object instance.

* Fix handling of request's body

We were not handling our request's body when receiving json. Fix tests
to catch that issue. Eventually we might want a test helper to remove
the boilerplate of sending json requests.

* Add missing lastname test to player model