Fix schema test issue
Im dumb, I tested the Migrator and Schema with the assumption that the
test database's schema version would be 0. Now that we have some
migrations, the test database's schema version will never be zero.

To fix that we create a decorator that cleans the schema version for
tests that needs it.