Create Migrator (#26)
* Move test database setup to fixtures module

Put all test fixtures in the `tests.fixtures` module for
discoverability.

* Implement Migrator.migrate_latest()

Start by implementing a naive migrate latest. This function currently
takes all of the migrations in a directory and performs them. Eventually
we only want to perform pending migrations.

* Store schema versions

We want to store the latest migration version to track the current state
of a given database and eventually know which migrations are pending and
needs to run.

* Make schema version table creation indempotent

Migrate latest will be run multiple times on a given database so we want
to make it idempotent. Start by fixing the schema_versions table
creation.

* Implement migrator.perform_migration

This will be useful when testing migrate latest and rollback (since
we'll be able to execute 1 migration then check if we can migrate or
rollback from it).

* Make migrate_latest idempotent

When some migrations have already been performed we only perform the
pending ones. We check the current schema version to figure out which
migrations are pending.

* Extract Schema-related logic from Migrator

Create a Schema object to cleanup the all of the version logic away from
the Migrator.