Create migration script (#28)
* Create first migration

Kickstart our migrations by creating the players table.

* Create development migration script

We might pass a config argument to that script to reuse the logic for
different environment.

* Parse string to configuration

This will be useful when parsing commandline argument to configuration
objects in our migrate latest script

* Create models folders in unit tests

Mimic the source code structure by moving the models tests in a
designated folder.

* Log performed migrations

When performing the migration script we output the number of performed
migrations and the environment for debugging.

* Add migration instruction to readme

Document the migration script in the readme. Eventually we might have a
documentation index with a designated `/docs/database.md` but for now
that's overkill.

* Fix schema test issue

Im dumb, I tested the Migrator and Schema with the assumption that the
test database's schema version would be 0. Now that we have some
migrations, the test database's schema version will never be zero.

To fix that we create a decorator that cleans the schema version for
tests that needs it.

* Fix test warning

The prefix `test` seems to be reserved by pytest so we replace it with
`mock` for our spy object.

* Add migration path default argument

We only need to inject the path for testing purpose. Provide a sane
default when using the Migrator in our scripts.

* Handle singular migration output

When performing a single migration we use the singular `migration`
instead of `migrations`.

* Move schema.create_if_needed out of __init__

It's weird to have a side effect (creation of the schema table) in the
initializer of the migrator.

* Make config.for_key logic dynamic

Match the key to the beginning of the config's name instead of using a
big hardcoded switch.