Persist Players in Database (#29)
* Make Database's behaviors more intuitive

Instead of raising errors when the database is not open we either ignore
the message (on rollback/close) or open the database before executing a
sql query.

* Create controller baseclass

This will enable us to do dependency injection in our controllers. We
need that in order to inject our database object.

* Move tournament controller to our new class

* Create player repository

Create a repository object to store/fetch players from the database and
use it in our controller instead of faking the player's creation.

* Remove unnecessary pytest imports

* Add integration test for player creation

Make sure that the POST /players request persist the player in the
database.

* Infer controller's resource name from class name

Use the controller's class name to compute a resource name for our
route. This approach favors naming convention over configurations.