persist tournaments in database (#31)
* CREATE tournament table

* create participations table

we want to associate players to a tournament, and the problem is that both belong to each other multiple times, hence we need a table to keep track.

* Fix linterwarning in migration

changed gitignore because it wasn't ignoring vscode .json files
remigrated because I didnt have linter installed so I had a missing line break

* Persists participations in database

Due to psychopg limitations we need to do 2 queries. We will fix that in a future commit.

* Refactor participation repository

Clean up the participation repository by breaking create_many into
multiple private method and a build static function on the model.

* Expose find_by_tournament method

We'll use this when fetching all players for a given tournament so might
as well expose this from our repository's api.

* Delete empty tests

These test were covering a fake in-memory implementation of the
Tournament controller. We don't need them for now.

* fixup! create participations table

* fixup! CREATE tournament table

* fixup! Persists participations in database